+++ packages/kit/src/core/config/options.js
@@ -134,6 +134,10 @@ const options = object(
 					(keypath) => `${keypath} has been renamed to config.kit.moduleExtensions`
 				),
 
+				env: object({
+					publicPrefix: string('PUBLIC_')
+				}),
+
 				files: object({
 					assets: string('static'),
 					hooks: string(join('src', 'hooks')),